created navbar
