+++ fitly/app/src/main/java/com/example/a007fa/fitly/fitlyHandler.java
@@ -13,7 +13,6 @@ import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
 import android.support.v4.content.LocalBroadcastManager;
 
-import java.util.Calendar;
 import java.util.ArrayList;
 
 import android.os.IBinder;
@@ -76,20 +75,20 @@ public class fitlyHandler extends Service implements SensorEventListener {
 
 
     protected void sendStepMessage() {
-        Intent intent = new Intent(getApplicationContext(), Dashboard.class);
+        Intent intent = new Intent(getApplicationContext(), MainActivity.class);
         intent.setAction(ACTION_FITLY);
         intent.putExtra("stepCount", steps);
         LocalBroadcastManager.getInstance(getApplicationContext()).sendBroadcast(intent);
     }
 
     protected void sendBadgeMessage() {
-        Intent intent = new Intent(getApplicationContext(), Dashboard.class);
+        Intent intent = new Intent(getApplicationContext(), MainActivity.class);
         intent.setAction(ACTION_BADGE);
         LocalBroadcastManager.getInstance(getApplicationContext()).sendBroadcast(intent);
     }
 
     protected void sendBigBadgeMessage() {
-        Intent intent = new Intent(getApplicationContext(), Dashboard.class);
+        Intent intent = new Intent(getApplicationContext(), MainActivity.class);
         intent.setAction(ACTION_BIGBADGE);
         LocalBroadcastManager.getInstance(getApplicationContext()).sendBroadcast(intent);
     }
@@ -129,7 +128,7 @@ public class fitlyHandler extends Service implements SensorEventListener {
         @Override
         public void onReceive(Context context, Intent intent) {
             if (intent.getAction().equals(ACTION_ENDDAY)) {
-                Intent intent1 = new Intent(getApplicationContext(), Dashboard.class);
+                Intent intent1 = new Intent(getApplicationContext(), MainActivity.class);
                 intent1.setAction(ACTION_FITLY);
                 intent1.putExtra("stepCount", steps);
                 bManager.sendBroadcast(intent);
