+++ fitly/app/src/main/java/com/example/a007fa/fitly/saveAlarm.java
@@ -1,7 +1,5 @@
 package com.example.a007fa.fitly;
 
-import android.content.BroadcastReceiver;
-import android.content.Context;
 import android.content.Intent;
 import android.os.IBinder;
 import android.support.v4.content.LocalBroadcastManager;
@@ -13,7 +11,7 @@ public class saveAlarm extends Service {
     static final String ACTION_ENDDAY = "com.fitly.action.ENDDAY";
 
     public void onCreate() {
-        Intent intent1 = new Intent(getApplicationContext(), Dashboard.class);
+        Intent intent1 = new Intent(getApplicationContext(), MainActivity.class);
         intent1.setAction(ACTION_ENDDAY);
         LocalBroadcastManager.getInstance(getApplicationContext()).sendBroadcast(intent1);
     }
