+++ fitly/app/src/main/res/color/nav_item_colors.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/colorWhite"
+        android:state_checked="true" />
+    <item android:color="@color/colorDarkened" />
+</selector>
