+++ fitly/app/src/main/res/values/colors.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+
     <color name="colorPrimary">#77bed5</color>
     <color name="colorPrimaryDark">#1c4b5b</color>
     <color name="colorAccent">#e2ce62</color>
+    <color name="colorWhite">#ffffff</color>
+    <color name="colorDarkened">#c7e3ed</color>
 </resources>
