+++ fitly/app/src/main/res/values/strings.xml
@@ -10,4 +10,7 @@
     <string name="title_badges">Badges</string>
     <string name="title_dashboard">Dashboard</string>
     <string name="title_profile">Profile</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
